Tweaked scripts

+++ mobile/linearmqtt/reset-device-count.sh
@@ -1,1 +1,3 @@
-rm devicecount.txt
+if [ -f "devicecount.txt" ]; then
+  rm devicecount.txt
+fi
